Prompt Sensitivity Score/pyproject.toml
[project]
name = "pss-pipeline"
version = "1.0.0"
description = "Prompt Sensitivity Score pipeline for tiny-aya-global"
requires-python = ">=3.9"

dependencies = [
    "cohere>=5.20.0",
    "python-dotenv>=1.0.0",
    "datasets>=3.0.0",
    "sentence-transformers>=3.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "transformers>=4.40.0",
]

[project.optional-dependencies]
nlp = [
    "spacy>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "analysis", "utils", "config", "plots"]
